% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{get_nba_results}
\alias{get_nba_results}
\title{NBA results for a day}
\usage{
get_nba_results(game_date)
}
\arguments{
\item{game_date}{date or character format "\%Y-\%m-\%d"}
}
\value{
results for a day
}
\description{
NBA results for a day
}
