% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_scores.R
\name{get_nba_scores}
\alias{get_nba_scores}
\title{All NBA scores for a day}
\usage{
get_nba_scores(game_date)
}
\arguments{
\item{game_date}{date or character format "\%Y-\%m-\%d"}
}
\value{
scores for a day
}
\description{
All NBA scores for a day
}
\examples{
get_nba_scores("2023-03-21")
}
