% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winamax.R
\name{get_odds}
\alias{get_odds}
\title{Get odds from bet id}
\usage{
get_odds(json, bet_id)
}
\arguments{
\item{json}{json file containing the desired odds according to the bet id.}

\item{bet_id}{MainbetId}
}
\value{
Odds
}
\description{
Get odds from bet id
}
