% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winamax.R
\name{get_sport_df}
\alias{get_sport_df}
\title{Get all odds for selected sports}
\usage{
get_sport_df(sport, bet_status, next_hours = 24)
}
\arguments{
\item{sport}{Sports list between "Tennis", "Basketball", and "Baseball".
Others sports are not yet implemented.}

\item{bet_status}{"PREMATCH" or "LIVE"}

\item{next_hours}{next_hours}
}
\value{
Odds for selected sports.
}
\description{
Get all odds for selected sports
}
\examples{
library(rvest)
library(xml2)
library(jsonlite)

# Tennis bets for next 3 hours:
get_sport_df(sport = "Tennis", bet_status = "PREMATCH", next_hours = 3)
}
