% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winamax.R
\name{get_winamax_json}
\alias{get_winamax_json}
\title{Get json in winamax source code}
\usage{
get_winamax_json(url, ...)
}
\arguments{
\item{url}{winamax url}

\item{...}{other options to implement (categoryId, etc...)}
}
\value{
json file
}
\description{
Get json in winamax source code
}
\examples{
library(rvest)
library(xml2)
library(jsonlite)

json <- get_winamax_json("https://www.winamax.fr/paris-sportifs/sports/2/15/177")
}
