% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SportsIdName}
\alias{SportsIdName}
\title{Sports Id on Winamax}
\format{
\subsection{\code{SportsIdName}}{

A data frame with 22 Id for sports on Winamax:
\describe{
\item{SportId}{Sport Id on Winamax}
\item{SportName}{Sport name}
...
}
}
}
\source{
\url{https://www.winamax.fr/paris-sportifs}
}
\usage{
SportsIdName
}
\description{
On data frame with Id for all Winamax sports
}
\keyword{datasets}
